feat: disable Approve button when task checks are not ready
